% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_px_file.R
\name{check_file_or_url}
\alias{check_file_or_url}
\title{Check download url whether it is a px cube}
\usage{
check_file_or_url(file_or_url)
}
\arguments{
\item{px_download_url}{url to download a px cube}
}
\value{
file_is_px_axis_file
}
\description{
Check download url whether it is a px cube
}
\examples{
check_px_cube_url('https://www.pxweb.bfs.admin.ch/DownloadFile.aspx?file=px-x-0102020207_102')
}
