% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file.R
\name{scan_px_file}
\alias{scan_px_file}
\title{Scan a px cube file and return px data and metadata}
\usage{
scan_px_file(file_or_url)
}
\arguments{
\item{px_file}{px cube file}
}
\value{
list with the dataframe as a tibble and the metadata as a list
}
\description{
Scan a px cube file and return px data and metadata
}
\examples{
scan_px_file("px-x-0102020203_110.px")
}
